Add language attribute to code-language blocks.
Currently, this is kind of redundant.  The language labels are present both
on the headers rendered in normal Markdown and in the block attributes.  At
some point, perhaps we could "scrape" the attribute value out of the HTML
similar to how the "source" key on attribute definitions functions in JS.
Might not be worth the complexity, though.